scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
python-dotenv==1.0.0
pymongo==4.4.1
lxml==4.9.3
APScheduler==3.10.4
